// ddg-frontend/src/services/openapi/models/PdfSignatory_pdf_write.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PdfSignatory_pdf_write = {
    /**
     * The signatory name for the center that will be used in the signature page
     */
    centerName?: string | null;
    /**
     * The signatory name for the company that will be used in the signature page
     */
    companyName?: string | null;
    /**
     * The second signatory name for the company that will be used in the signature page
     */
    companyName2?: string | null;
    place?: string | null;
    /**
     * The signatory date that will be used in the signature page
     */
    date?: string | null;
};